Day_3/requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.14
langchain-core>=0.2.38
langchain-community>=0.2.10
langgraph>=0.0.64
langchain-google-genai>=0.0.12
langchain-text-splitters>=0.2.2
typing-extensions>=4.12.2
pypdf>=4.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
streamlit>=1.37.0
nest-asyncio>=1.5.8
